+++ apps/neoscan/priv/repo/migrations/20180607000003_vouts.exs
@@ -8,7 +8,7 @@ defmodule Neoscan.Repo.Migrations.Vouts do
       add(:n, :integer, null: false, primary_key: true)
       add(:address_hash, :binary, null: false)
       add(:asset_hash, :binary, null: false)
-      add(:value, :float, null: false)
+      add(:value, :decimal, null: false)
       add(:block_time, :naive_datetime, null: false)
 
       add(:claimed, :boolean, null: false, default: false)
